Fix sibling name check for creation and editing
