+++ src/containers/MyNdla/Folders/EditFolderModal.tsx
@@ -8,9 +8,15 @@
 
 import { ModalBody, ModalCloseButton, ModalHeader, ModalV2 } from '@ndla/modal';
 import { useTranslation } from 'react-i18next';
+import { useMemo } from 'react';
+import { useApolloClient } from '@apollo/client';
 import { GQLFolder } from '../../../graphqlTypes';
 import FolderForm from './FolderForm';
-import { useUpdateFolderMutation } from '../folderMutations';
+import {
+  getFolder,
+  useFolders,
+  useUpdateFolderMutation,
+} from '../folderMutations';
 
 interface Props {
   folder?: GQLFolder;
@@ -23,6 +29,20 @@ const EditFolderModal = ({ folder, isOpen, onClose, onSaved }: Props) => {
   const { t } = useTranslation();
 
   const { updateFolder, loading } = useUpdateFolderMutation();
+
+  const { cache } = useApolloClient();
+
+  const { folders } = useFolders();
+
+  const levelFolders = useMemo(
+    () =>
+      folder?.parentId
+        ? getFolder(cache, folder.parentId)?.subfolders ?? []
+        : folders,
+    [cache, folder?.parentId, folders],
+  );
+
+  const siblings = levelFolders.filter((f) => f.id !== folder?.id);
 
   return (
     <ModalV2
@@ -44,6 +64,7 @@ const EditFolderModal = ({ folder, isOpen, onClose, onSaved }: Props) => {
             {folder && (
               <FolderForm
                 folder={folder}
+                siblings={siblings}
                 onSave={async (values) => {
                   await updateFolder({
                     variables: {
